+++ packages/moonstone/CHANGELOG.md
@@ -8,6 +8,8 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 - `moonstone/VideoPlayer` space allotment on both sides of the playback controls to support 3 buttons; consequently the "more" controls area has shrunk by the same amount
 - `moonstone/Marquee` property `rtl` to set directionality to right-to-left
+- `moonstone/FormCheckboxItem`
+- `moonstone/FormCheckbox`, a standalone checkbox, to support `moonstone/FormCheckboxItem`
 
 ### Fixed